Convert prompt to .vue component
