+++ assets/js/main.js
@@ -7,7 +7,7 @@ raven.init();
 
 common.init();
 
-const vueSplit = () => import(/* webpackChunkName: "vue-components" */ './bootstraps/vue');
+const vueSplit = () => import(/* webpackChunkName: "vue-components" */ './vue-apps/index');
 vueSplit().then(vueComponents => {
     vueComponents.init();
 });
